first commit, set up the basic file structure with the frontend and the backend
the backend firstly has the package.json, which i added using the npm init, set up the models folder, which will contain all the types of collections and their schema, for now only User.js is there but will add Post.js later and update User.js, wrote the code for connection of mongodb to the server in server.js. wrote the logic for adding new users, getting the list of users and getting user by id inside the users.js file of the routes folder.